# Send frame diffs over instead of full frames