--- a/include/state.h
+++ b/include/state.h
@@ -12,6 +12,7 @@
 #include <set>
 #include <vector>
 
+#include "BWEnv/fbs/messages_generated.h"
 #include "constants.h"
 #include "frame.h"
 #include "refcount.h"
@@ -156,6 +157,7 @@ class State : public RefCounted {
   bool setRawImage(const torchcraft::fbs::Frame* frame);
   void postUpdate(std::vector<std::string>& upd);
   bool checkBattleFinished(std::vector<std::string>& upd);
+  void update_frame(const torchcraft::fbs::FrameData* fd);
 
   bool microBattles_;
   std::set<BW::UnitType> onlyConsiderTypes_;