Now the connection also uses frame diffs instead of the full frames.

I ran clang-format on every file this change touched, so a lot of this is automatic, sorry guys :( I'll split it into 2 commits next time. module.cc wasn't really changed much except for the destructor. 

In frame.h I removed all of the diff functions involving references, so we don't get an explosion of signatures that take either pointers or references,