@ebetica 2017-06-14
Timing until [this line](https://github.com/TorchCraft/TorchCraft/blob/develop/examples/dump_replay.lua#L55) on bwrep_gwhxe.rep for the first 5000 frames:

With diffs:
th dump_replay.lua -t localhost  11.05s user 0.58s system 44% cpu 26.142 total

With diffs but without reserializing the frame string:
th dump_replay.lua -t localhost  3.47s user 0.54s system 21% cpu 18.475 total

Without diffs:
th dump_replay.lua -t localhost  10.80s user 0.63s system 35% cpu 32.410 total

What do we need the [frame string](https://github.com/TorchCraft/TorchCraft/pull/167/files#diff-4c76f9fac82dd94513c6286362f89c33R216) for and can I remove it @syhw @jgehring